Rename rams js core file from index.js to RAMS.js
